put User & Order models into /model directory
