+++ Go-Backend/model/Order.go
@@ -1,0 +1,11 @@
+package model
+
+import "gorm.io/gorm"
+
+type Order struct {
+	gorm.Model
+	ID    uint `gorm:"primaryKey;unique"`
+	Name  string
+	Desc  string
+	Image string
+}
