+++ Go-Backend/model/User.go
@@ -1,0 +1,10 @@
+package model
+
+import "gorm.io/gorm"
+
+type User struct {
+	gorm.Model
+	ID   uint `gorm:"primaryKey;unique"`
+	Name string
+	Age  uint
+}
